+++ templates/repo/issue/list.tmpl
@@ -295,7 +295,7 @@
 						</span>
 						<div class="menu">
 							<div class="item issue-action" data-action="clear" data-url="{{$.RepoLink}}/issues/labels">
-								aaa{{.locale.Tr "repo.issues.new.clear_labels"}}
+								{{.locale.Tr "repo.issues.new.clear_labels"}}
 							</div>
 							{{$previousExclusiveScope := "_no_scope"}}
 							{{range .Labels}}
